MQTTManager/requirements.txt
flask
paho-mqtt
fastapi[all]
uvicorn
python-jose[cryptography]
passlib[bcrypt]